✨ Feat: Project API 통합 및 참여 로직 변경
